Introduce a MOJO_TEMPLATE_COVER environment variable
When this environment variable is set to a true value, the anonymous sub that
is generated from the template, is injected into the symbol table.

Making the anonymous sub accessible from the symbol table, allows Devel::Cover
to see it and report correct coverage for it.

However storing the sub in the symbol table also creates a memory leak, which
is why this feature is only enabled when the MOJO_TEMPLATE_COVER is set to a
true value.